Added gsl_rng functions to pfmodule.c
Made sure they all started with gsl_rng, so that meant that old
pf.get_gsl_rng() became pf.gsl_ran_get().  It was used in data.py and
func.py, so they were changed.

A get/set pair of functions allows getting the state of the rng to
save and restore later.