Added Mcmc.blankSeqNums, to inform and restrict compo simulations
This is a list of lists of sequence numbers, one list for each data
partition, where the sequence is completely blank.  Since these
sequences do not contribute to the bigXSq for the original data, they
should not contribute to the bigXSq of simulations for posterior
predictive simulations, and so contributions to bigXSq from those
sequences are excluded.  See the doc string for
Data.compoChiSquaredTest()